Added custom Logs using Logger

+++ src/tasks/tasks.service.ts
@@ -1,15 +1,16 @@
 import { User } from './../auth/user.entity';
-import { EntityManager, Repository } from 'typeorm';
+import { Repository } from 'typeorm';
 import { GetTaskFilterDTO } from './dto/get-task-filter.dto';
 import { CreateDTO } from './dto/create-task.dto';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, Logger, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
-import { TaskRepository } from './task.repository';
+/* import { TaskRepository } from './task.repository'; */
 import { TaskStatus } from './task-status.enum';
 
 @Injectable()
 export class TasksService {
+    private logger = new Logger("TaskService")
     constructor( 
         @InjectRepository(Task)
         private taskRepository: Repository<Task>,
@@ -32,8 +33,15 @@ export class TasksService {
             { search: `%${search}%` })
         }
 
-        const tasks = await query.getMany()
-        return tasks
+        try {
+            const tasks = await query.getMany()
+            return tasks
+        } catch (error) {
+            this.logger.error(`Failed to get tasks for user ${user.username}. Filters: ${JSON.stringify(filterDTO)}`, error.stack)
+            throw new InternalServerErrorException()
+        }
+
+        
     }
 
     async createTask(
@@ -47,14 +55,23 @@ export class TasksService {
         task.description = description
         task.status = TaskStatus.OPEN
         task.user = user
-        await task.save()
+
+        try {
+            await task.save()
+        } catch (error) {
+            this.logger.error(`Failed to create a task for user ${user.username}. Data: ${JSON.stringify(createDTO)}`,error.stack)
+            throw new InternalServerErrorException()
+        }
         delete task.user
 
         return task
     }  
 
-    getTaskById(id: number): Promise<Task> {
-        const found = this.taskRepository.findOneBy({ id });
+    getTaskById(
+        id: number,
+        user: User
+    ): Promise<Task> {
+        const found = this.taskRepository.findOne({ where: {id, userId: user.id} });
         if(!found) {
             throw new NotFoundException(`Task with ID ${id} not found!`)
         } else {
@@ -62,16 +79,23 @@ export class TasksService {
         }
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id)
+    async deleteTask(
+        id: number,
+        user: User,
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id })
         
         if(result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found!`)
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id)
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user)
         task.status = status
         await task.save()
         return task
